--- cmake/after_load_conf.cmake
@@ -1,6 +1,0 @@
-#
-#
-#
-
-# Add definitions
-add_definitions(-DMOD_AH_BOT)
