+++ src/AuctionHouseBot.cpp
@@ -23,7 +23,6 @@
 #include "Config.h"
 #include "Player.h"
 #include "WorldSession.h"
-#include "ScriptMgr.h"
 #include <vector>
 
 using namespace std;
@@ -1885,81 +1884,4 @@ void AuctionHouseBot::LoadValues(AHBConfig *config)
         }
     }
 	if (debug_Out) sLog->outError( "End Settings for %s Auctionhouses:", WorldDatabase.PQuery("SELECT name FROM auctionhousebot WHERE auctionhouse = %u", config->GetAHID())->Fetch()->GetCString());
-}
-
-class AHBot_WorldScript : public WorldScript
-{
-public:
-    AHBot_WorldScript() : WorldScript("AHBot_WorldScript") { }
-
-    void OnBeforeConfigLoad(bool /*reload*/) override
-    {
-        auctionbot->InitializeConfiguration();
-    }
-
-    void OnLoadCustomDatabaseTable() override
-    {
-        sLog->outString("Initialize AuctionHouseBot...");
-        auctionbot->Initialize();
-    }
-};
-
-class AHBot_AuctionHouseScript : public AuctionHouseScript
-{
-public:
-    AHBot_AuctionHouseScript() : AuctionHouseScript("AHBot_AuctionHouseScript") { }
-
-    void OnBeforeAuctionHouseMgrSendAuctionSuccessfulMail(AuctionHouseMgr* /*auctionHouseMgr*/, AuctionEntry* /*auction*/, Player* owner, uint32& /*owner_accId*/, uint32& /*profit*/, bool& sendNotification, bool& updateAchievementCriteria) override
-    {
-        if (owner && owner->GetGUIDLow() == auctionbot->GetAHBplayerGUID())
-        {
-            sendNotification = false;
-            updateAchievementCriteria = false;
-        }
-    }
-
-    void OnBeforeAuctionHouseMgrSendAuctionExpiredMail(AuctionHouseMgr* /*auctionHouseMgr*/, AuctionEntry* /*auction*/, Player* owner, uint32& /*owner_accId*/, bool& sendNotification) override
-    {
-        if (owner && owner->GetGUIDLow() == auctionbot->GetAHBplayerGUID())
-            sendNotification = false;
-    }
-
-    void OnBeforeAuctionHouseMgrSendAuctionOutbiddedMail(AuctionHouseMgr* /*auctionHouseMgr*/, AuctionEntry* auction, Player* oldBidder, uint32& /*oldBidder_accId*/, Player* newBidder, uint32& newPrice, bool& /*sendNotification*/) override
-    {
-        if (oldBidder && !newBidder)
-            oldBidder->GetSession()->SendAuctionBidderNotification(auction->GetHouseId(), auction->Id, auctionbot->GetAHBplayerGUID(), newPrice, auction->GetAuctionOutBid(), auction->item_template);
-    }
-
-    void OnAuctionAdd(AuctionHouseObject* /*ah*/, AuctionEntry* auction) override
-    {
-        auctionbot->IncrementItemCounts(auction);
-    }
-
-    void OnAuctionRemove(AuctionHouseObject* /*ah*/, AuctionEntry* auction) override
-    {
-        auctionbot->DecrementItemCounts(auction, auction->item_template);
-    }
-};
-
-class AHBot_MailScript : public MailScript
-{
-public:
-    AHBot_MailScript() : MailScript("AHBot_MailScript") { }
-
-    void OnBeforeMailDraftSendMailTo(MailDraft* /*mailDraft*/, MailReceiver const& receiver, MailSender const& sender, MailCheckMask& /*checked*/, uint32& /*deliver_delay*/, uint32& /*custom_expiration*/, bool& deleteMailItemsFromDB, bool& sendMail) override
-    {
-        if (receiver.GetPlayerGUIDLow() == auctionbot->GetAHBplayerGUID())
-        {
-            if (sender.GetMailMessageType() == MAIL_AUCTION)        // auction mail with items
-                deleteMailItemsFromDB = true;
-            sendMail = false;
-        }
-    }
-};
-
-void AddAHBotScripts()
-{
-    new AHBot_WorldScript();
-    new AHBot_AuctionHouseScript();
-    new AHBot_MailScript();
 }
