+++ src/loader_ah_bot.h
@@ -1,0 +1,1 @@
+void AddAHBotScripts();
