+++ src/loader_cs_ah_bot.h
@@ -1,7 +1,1 @@
-void AddSC_ah_bot_commandscript();
-void AddAHBotScripts();
-
-void AddAHBotCommandScripts()
-{
-    AddSC_ah_bot_commandscript();
-}
+void AddAHBotCommandScripts();
